Seperate MMU from LSU
