+++ core/CPUFactory.cpp
@@ -90,8 +90,10 @@ auto olympia::CPUFactory::bindTree_(sparta::RootTreeNode* root_node,
         auto core_tree_node = root_node->getChild(std::string("cpu.core") +
                                                   std::to_string(num_of_cores));
         sparta_assert(core_tree_node != nullptr);
+        (core_tree_node->getChild("mmu")->getResourceAs<olympia::MMU>())->
+                setTLB(*private_nodes_.at(num_of_cores)->getResourceAs<olympia::SimpleTLB>());
         (core_tree_node->getChild("lsu")->getResourceAs<olympia::LSU>())->
-            setTLB(*private_nodes_.at(num_of_cores)->getResourceAs<olympia::SimpleTLB>());
+                setMMU(*core_tree_node->getChild("mmu")->getResourceAs<olympia::MMU>());
         (core_tree_node->getChild("preloader")->getResourceAs<olympia::Preloader>())->
             preload();
     }
