+++ core/Cache.hpp
@@ -1,0 +1,14 @@
+//
+// Created by suraj on 8/24/23.
+//
+
+#ifndef OLYMPIA_CACHE_H
+#define OLYMPIA_CACHE_H
+
+
+class Cache {
+
+};
+
+
+#endif //OLYMPIA_CACHE_H
